Fix return type of function and add 2 tests
